Fix next page button not working
