+++ buildsystem-abstraction/adapter-1_13/src/main/java/com/eintosti/buildsystem/version/v1_13_R1/GameRules_1_13_R1.java
@@ -8,13 +8,11 @@
 
 package com.eintosti.buildsystem.version.v1_13_R1;
 
-import com.eintosti.buildsystem.version.GameRules;
-import org.bukkit.Bukkit;
+import com.eintosti.buildsystem.version.gamerules.AbstractGameRulesInventory;
 import org.bukkit.ChatColor;
 import org.bukkit.GameRule;
 import org.bukkit.Material;
 import org.bukkit.World;
-import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemFlag;
@@ -22,83 +20,29 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.UUID;
 
 /**
  * @author einTosti
  */
-public class GameRules_1_13_R1 implements GameRules {
+public class GameRules_1_13_R1 extends AbstractGameRulesInventory {
 
-    private final Map<UUID, Integer> invIndex;
-    private final int[] slots = new int[]{11, 12, 13, 14, 15, 20, 21, 22, 23, 24, 29, 30, 31, 32, 33};
-
-    private final String inventoryTitle;
     private final List<String> booleanEnabledLore, booleanDisabledLore;
-    private final List<String> unknownLore;
     private final List<String> integerLore;
 
-    private Inventory[] inventories;
-    private int numGameRules = 0;
+    public GameRules_1_13_R1(String inventoryTitle, List<String> booleanEnabledLore, List<String> booleanDisabledLore, List<String> ignoredUnknownLore, List<String> integerLore) {
+        super(inventoryTitle);
 
-    public GameRules_1_13_R1(String inventoryTitle, List<String> booleanEnabledLore, List<String> booleanDisabledLore, List<String> unknownLore, List<String> integerLore) {
-        this.invIndex = new HashMap<>();
-
-        this.inventoryTitle = inventoryTitle;
         this.booleanEnabledLore = booleanEnabledLore;
         this.booleanDisabledLore = booleanDisabledLore;
-        this.unknownLore = unknownLore;
         this.integerLore = integerLore;
     }
 
-    private Inventory createInventory() {
-        return Bukkit.createInventory(null, 45, inventoryTitle);
-    }
-
     @Override
-    public Inventory getInventory(Player player, World world) {
-        addGameRules(world);
-        return inventories[getInvIndex(player.getUniqueId())];
-    }
-
-    @Override
-    public void addGameRules(World world) {
-        String[] gameRules = world.getGameRules();
-
-        int columnGameRule = 0, maxColumnGameRule = 14;
-        setNumGameRules(world);
-        int numInventories = (numGameRules % 15 == 0 ? numGameRules : numGameRules + 1) != 0 ? (numGameRules % 15 == 0 ? numGameRules : numGameRules + 1) : 1;
-
-        inventories = new Inventory[numInventories];
-        Inventory inventory = createInventory();
-
-        int index = 0;
-        inventories[index] = inventory;
-
-        for (int i = 0; i < gameRules.length; i++) {
-            addGameRuleItem(inventory, this.slots[columnGameRule++], world, world.getGameRules()[i]);
-            if (columnGameRule > maxColumnGameRule) {
-                columnGameRule = 0;
-                inventory = createInventory();
-                inventories[++index] = inventory;
-            }
-        }
-    }
-
-    private boolean isBoolean(String gameRuleName) {
-        GameRule<?> gameRule = GameRule.getByName(gameRuleName);
-        if (gameRule == null) {
-            return false;
-        }
-        return gameRule.getType().equals(Boolean.class);
-    }
-
-    private void addGameRuleItem(Inventory inventory, int slot, World world, String gameRule) {
+    protected void addGameRuleItem(Inventory inventory, int slot, World world, String gameRule) {
         ItemStack itemStack = new ItemStack(isEnabled(world, gameRule) ? Material.FILLED_MAP : Material.MAP);
-
         ItemMeta itemMeta = itemStack.getItemMeta();
+
         itemMeta.setDisplayName("§e" + gameRule);
         itemMeta.setLore(getLore(world, gameRule));
         itemMeta.addItemFlags(ItemFlag.values());
@@ -107,7 +51,8 @@ public class GameRules_1_13_R1 implements GameRules {
         inventory.setItem(slot, itemStack);
     }
 
-    private boolean isEnabled(World world, String gameRuleName) {
+    @Override
+    protected boolean isEnabled(World world, String gameRuleName) {
         GameRule<?> gameRule = GameRule.getByName(gameRuleName);
         if (gameRule == null) {
             return false;
@@ -138,13 +83,12 @@ public class GameRules_1_13_R1 implements GameRules {
         return lore;
     }
 
-    private boolean isValidSlot(int slot) {
-        for (int i : this.slots) {
-            if (i == slot) {
-                return true;
-            }
+    private boolean isBoolean(String gameRuleName) {
+        GameRule<?> gameRule = GameRule.getByName(gameRuleName);
+        if (gameRule == null) {
+            return false;
         }
-        return false;
+        return gameRule.getType().equals(Boolean.class);
     }
 
     @Override
@@ -160,16 +104,13 @@ public class GameRules_1_13_R1 implements GameRules {
         }
 
         ItemMeta itemMeta = itemStack.getItemMeta();
-        if (itemMeta == null) {
+        if (itemMeta == null || !itemMeta.hasDisplayName()) {
             return;
         }
 
-        if (!itemMeta.hasDisplayName()) {
-            return;
-        }
         String displayName = itemMeta.getDisplayName();
-
         String gameRuleName = ChatColor.stripColor(displayName);
+
         if (isBoolean(gameRuleName)) {
             GameRule<Boolean> gameRule = (GameRule<Boolean>) GameRule.getByName(gameRuleName);
             Boolean value = world.getGameRuleValue(gameRule);
@@ -194,46 +135,5 @@ public class GameRules_1_13_R1 implements GameRules {
 
             world.setGameRule(gameRule, value);
         }
-    }
-
-    @Override
-    public void incrementInv(Player player) {
-        UUID playerUUID = player.getUniqueId();
-        invIndex.put(playerUUID, invIndex.get(playerUUID) + 1);
-    }
-
-    @Override
-    public void decrementInv(Player player) {
-        UUID playerUUID = player.getUniqueId();
-        invIndex.put(playerUUID, invIndex.get(playerUUID) - 1);
-    }
-
-    @Override
-    public int getInvIndex(UUID uuid) {
-        if (!invIndex.containsKey(uuid)) {
-            resetInvIndex(uuid);
-        }
-        return invIndex.get(uuid);
-    }
-
-    @Override
-    public void resetInvIndex(UUID uuid) {
-        invIndex.put(uuid, 0);
-    }
-
-    @Override
-    public int getNumGameRules() {
-        return numGameRules;
-    }
-
-    private void setNumGameRules(World world) {
-        String[] gameRules = world.getGameRules();
-        int gameRuleLength = gameRules.length;
-        this.numGameRules = (gameRuleLength / 15) + (gameRuleLength % 15 == 0 ? 0 : 1);
-    }
-
-    @Override
-    public int[] getSlots() {
-        return slots;
     }
 }
