+++ buildsystem-abstraction/common/build.gradle.kts
@@ -6,4 +6,5 @@ plugins {
 
 dependencies {
     compileOnly(libs.spigot)
+    compileOnly(libs.annotations)
 }
