+++ buildsystem-abstraction/common/src/main/java/com/eintosti/buildsystem/version/customblocks/CustomBlocks.java
@@ -6,13 +6,14 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.eintosti.buildsystem.version;
+package com.eintosti.buildsystem.version.customblocks;
 
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Player;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * @author einTosti
@@ -29,5 +30,5 @@ public interface CustomBlocks {
 
     void toggleIronDoor(PlayerInteractEvent event);
 
-    void rotate(Block block, Player player, BlockFace blockFace);
+    void rotate(Block block, Player player, @Nullable BlockFace blockFace);
 }
