+++ buildsystem-core/src/main/java/com/eintosti/buildsystem/BuildSystem.java
@@ -36,8 +36,8 @@ import com.eintosti.buildsystem.util.Messages;
 import com.eintosti.buildsystem.util.RBGUtils;
 import com.eintosti.buildsystem.util.SkullCache;
 import com.eintosti.buildsystem.util.external.UpdateChecker;
-import com.eintosti.buildsystem.version.CustomBlocks;
-import com.eintosti.buildsystem.version.GameRules;
+import com.eintosti.buildsystem.version.customblocks.CustomBlocks;
+import com.eintosti.buildsystem.version.gamerules.GameRules;
 import org.bstats.bukkit.Metrics;
 import org.bstats.charts.SimplePie;
 import org.bukkit.Bukkit;
