+++ gradle.properties
@@ -1,5 +1,5 @@
 group=com.eintosti
-version=2.20.1
+version=2.20.2
 org.gradle.daemon=true
 org.gradle.caching=true
 org.gradle.parallel=true
